backend/requirements.txt
Flask>=2.0
flask-cors>=3.0.0
python-dotenv>=0.21
psycopg2-binary>=2.9.0
google-generativeai>=0.8.0
